Moved main Corale codebase to new base namespace
