+++ Corale.Colore/Colore.csproj.DotSettings
