+++ Corale.Colore/Colore.nuspec
