+++ Corale.Colore/Properties/AssemblyInfo.cs
