+++ Corale.Colore/Settings.StyleCop
